Refactor quiet path list
